+++ MonsterChase/Engine/World/Public/World.h
@@ -10,7 +10,7 @@ namespace Engine
 
 		void AddGameObject(SmartPtr<GameObject>& i_Object);
 		void RemoveGameObject(WeakPtr<GameObject>& i_Object);
-		WeakPtr<GameObject> GetGameObject(const int& i_index);
+		WeakPtr<GameObject> GetGameObject(const size_t& i_index);
 		size_t GetNumGameObjects();
 		void Destroy();
